p3.4.advanced-di-ioc/myfunction/requirements.txt
markdown2==2.3.5
PyYaml
